Begin decoupling session.sagas from external apps
